Typo fix + refactor
