+++ src/Dotconst/Loader.php
@@ -133,40 +133,4 @@ class Loader
 
         return $config;
     }
-
-    /**
-     * @param $path
-     *
-     * @return string
-     */
-    public static function normalizePath($path)
-    {
-        if (empty($path)) {
-            return '';
-        }
-
-        $path = str_replace(DIRECTORY_SEPARATOR, '/', $path);
-
-        $parts = explode('/', $path);
-
-        $safe = [];
-        foreach ($parts as $idx => $part) {
-            if (($idx == 0 && empty($part))) {
-                $safe[] = '';
-            } elseif (trim($part) == "" || $part == '.') {
-            } elseif ('..' == $part) {
-                if (null === array_pop($safe) || empty($safe)) {
-                    $safe[] = '';
-                }
-            } else {
-                $safe[] = $part;
-            }
-        }
-
-        if (count($safe) === 1 && $safe[0] === '') {
-            return DIRECTORY_SEPARATOR;
-        }
-
-        return implode(DIRECTORY_SEPARATOR, $safe);
-    }
 }
